CY-15: Crear serivicio para listar solicitudes por filtro y paginación

+++ domain/model/src/main/java/co/com/pragma/model/loantype/LoanType.java
@@ -3,16 +3,14 @@ package co.com.pragma.model.loantype;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 @Builder(toBuilder = true)
 public class LoanType {
 
     private Long id;
     private String name;
-    private BigDecimal minAmount;
-    private BigDecimal maxAmount;
+    private Double minAmount;
+    private Double maxAmount;
     private Float interestRate;
     private boolean automaticValidation;
